[DC-428] Add feedback context and to details page
