+++ src/pages/library/DataBrowserDetails.js
@@ -5,6 +5,7 @@ import { ButtonOutline, ButtonPrimary, ButtonSecondary, Link } from 'src/compone
 import FooterWrapper from 'src/components/FooterWrapper'
 import { centeredSpinner, icon } from 'src/components/icons'
 import { libraryTopMatter } from 'src/components/library-common'
+import Modal from 'src/components/Modal'
 import { ReactComponent as AzureLogo } from 'src/images/azure.svg'
 import { ReactComponent as GcpLogo } from 'src/images/gcp.svg'
 import { Ajax } from 'src/libs/ajax'
@@ -16,6 +17,7 @@ import { authStore } from 'src/libs/state'
 import * as Utils from 'src/libs/utils'
 import { commonStyles } from 'src/pages/library/common'
 import { datasetAccessTypes, importDataToWorkspace, uiMessaging, useDataCatalog } from 'src/pages/library/dataBrowser-utils'
+import { DataBrowserFeedbackModal } from 'src/pages/library/DataBrowserFeedbackModal'
 import { RequestDatasetAccessModal } from 'src/pages/library/RequestDatasetAccessModal'
 
 
@@ -109,6 +111,9 @@ const MainContent = ({ dataObj }) => {
 export const SidebarComponent = ({ dataObj, id }) => {
   const { access } = dataObj
   const [showRequestAccessModal, setShowRequestAccessModal] = useState()
+  const [feedbackShowing, setFeedbackShowing] = useState()
+  const [thanksShowing, setThanksShowing] = useState()
+  const sidebarWidth = 230
 
   return h(Fragment, [
     div({ style: { ...styles.content, width: 300, flexShrink: 0, display: 'flex', flexDirection: 'column', alignItems: 'center' } }, [
@@ -178,7 +183,7 @@ export const SidebarComponent = ({ dataObj, id }) => {
       h(ButtonOutline, {
         disabled: dataObj.access !== datasetAccessTypes.GRANTED,
         tooltip: dataObj.access === datasetAccessTypes.GRANTED ? '' : uiMessaging.controlledFeature_tooltip,
-        style: { fontSize: 16, textTransform: 'none', height: 'unset', width: 230, marginTop: 20 },
+        style: { fontSize: 16, textTransform: 'none', height: 'unset', width: sidebarWidth, marginTop: 20 },
         onClick: () => {
           Ajax().Metrics.captureEvent(`${Events.catalogView}:previewData`, {
             id: dataObj.id,
@@ -195,7 +200,7 @@ export const SidebarComponent = ({ dataObj, id }) => {
       h(ButtonPrimary, {
         disabled: dataObj.access !== datasetAccessTypes.GRANTED,
         tooltip: dataObj.access === datasetAccessTypes.GRANTED ? '' : uiMessaging.controlledFeature_tooltip,
-        style: { fontSize: 16, textTransform: 'none', height: 'unset', width: 230, marginTop: 20 },
+        style: { fontSize: 16, textTransform: 'none', height: 'unset', width: sidebarWidth, marginTop: 20 },
         onClick: () => {
           Ajax().Metrics.captureEvent(`${Events.catalogWorkspaceLink}:detailsView`, {
             id,
@@ -203,8 +208,39 @@ export const SidebarComponent = ({ dataObj, id }) => {
           })
           importDataToWorkspace([dataObj])
         }
-      }, ['Link to a workspace'])
+      }, ['Link to a workspace']),
+      div({ style: { display: 'flex', width: sidebarWidth, marginTop: 20 } }, [
+        icon('talk-bubble', { size: 60, style: { width: 60, height: 45 } }),
+        div({ style: { marginLeft: 10, lineHeight: '1.3rem' } }, [
+          h(Link, {
+            onClick: () => setFeedbackShowing(true)
+          },
+          ['Provide feedback']),
+          ' on this dataset view'
+        ])
+      ])
     ]),
+    feedbackShowing && h(DataBrowserFeedbackModal, {
+      title: '',
+      onDismiss: () => setFeedbackShowing(false),
+      onSuccess: () => {
+        setFeedbackShowing(false)
+        setThanksShowing(true)
+      },
+      primaryQuestion: 'Is there anything missing or that you would like to see in this dataset view?',
+      sourcePage: 'catalog-details'
+    }),
+    thanksShowing && h(Modal, {
+      onDismiss: () => setThanksShowing(false),
+      showCancel: false,
+      okButton: h(ButtonPrimary, {
+        onClick: () => setThanksShowing(false)
+      },
+      ['OK'])
+    },
+    [div({ style: { fontWeight: 600, fontSize: 18 } },
+      'Thank you for helping us improve the Data Catalog experience!')]
+    ),
     showRequestAccessModal && h(RequestDatasetAccessModal, {
       datasets: [dataObj],
       onDismiss: () => setShowRequestAccessModal(false)
