+++ src/pages/library/DataBrowserToggler.js
@@ -1,78 +1,15 @@
-import _ from 'lodash/fp'
-import { Fragment, useState } from 'react'
+import { useState } from 'react'
 import { div, h, label, strong } from 'react-hyperscript-helpers'
-import { ButtonPrimary, IdContainer, Link, spinnerOverlay, Switch } from 'src/components/common'
+import { ButtonPrimary, Link, Switch } from 'src/components/common'
 import { icon } from 'src/components/icons'
-import { TextArea, TextInput } from 'src/components/input'
 import Modal from 'src/components/Modal'
-import TooltipTrigger from 'src/components/TooltipTrigger'
-import { Ajax } from 'src/libs/ajax'
 import colors from 'src/libs/colors'
 import { isDataBrowserVisible } from 'src/libs/config'
-import { FormLabel } from 'src/libs/forms'
 import * as Nav from 'src/libs/nav'
 import { useStore } from 'src/libs/react-utils'
 import { authStore } from 'src/libs/state'
-import * as Utils from 'src/libs/utils'
 import { catalogPreviewStore } from 'src/pages/library/dataBrowser-utils'
-import validate from 'validate.js'
-
-
-const DataBrowserFeedbackModal = ({ onSuccess, onDismiss }) => {
-  const [contactEmail, setContactEmail] = useState('')
-  const [feedback, setFeedback] = useState('')
-  const [submitting, setSubmitting] = useState(false)
-
-  const submit = Utils.withBusyState(setSubmitting, async () => {
-    await Ajax().Surveys.submitForm('1FAIpQLSevEVLKiLNACAsti8k2U8EVKGHmQ4pJ8_643MfdY2lZEIusyw',
-      { 'entry.477992521': feedback, 'entry.82175827': contactEmail })
-    onSuccess()
-  })
-
-  const errors = validate({ feedback, contactEmail },
-    { feedback: { length: { minimum: 1, maximum: 2000 } }, contactEmail: contactEmail ? { email: true } : {} })
-
-  return h(Modal, {
-    onDismiss,
-    shouldCloseOnEsc: false,
-    width: 500,
-    title: 'Give feedback',
-    okButton: h(TooltipTrigger, { content: !!errors && _.map(error => div({ key: error }, [error]), errors) },
-      [h(ButtonPrimary, {
-        disabled: errors,
-        onClick: submit
-      }, ['Submit'])]
-    )
-  }, [
-    h(IdContainer, [id => h(Fragment, [
-      h(FormLabel, { required: true, htmlFor: id, style: { fontSize: 14, fontWeight: 300 } },
-        ['Please tell us about your experience with the new Data Catalog']),
-      h(TextArea, {
-        id,
-        autoFocus: true,
-        value: feedback,
-        onChange: setFeedback,
-        placeholder: 'Enter feedback',
-        style: { height: '8rem', marginTop: '0.25rem' }
-      }
-      )
-    ])]),
-    div({ style: { display: 'flex', justifyContent: 'flex-end', fontSize: 12 } }, ['2000 Character limit']),
-    h(IdContainer, [id => h(Fragment, [
-      h(FormLabel, { htmlFor: id, style: { fontSize: 14 } },
-        ['Can we contact you with further questions?']),
-      h(TextInput, {
-        id,
-        value: contactEmail,
-        onChange: setContactEmail,
-        placeholder: 'Enter email address',
-        style: { marginTop: '0.25rem' }
-      }
-      )
-    ])]),
-    submitting && spinnerOverlay
-  ])
-}
+import { DataBrowserFeedbackModal } from 'src/pages/library/DataBrowserFeedbackModal'
 
 
 export const DataBrowserPreviewToggler = ({ checked }) => {
@@ -129,7 +66,9 @@ export const DataBrowserPreviewToggler = ({ checked }) => {
       onSuccess: () => {
         setFeedbackShowing(false)
         setThanksShowing(true)
-      }
+      },
+      primaryQuestion: 'Please tell us about your experience with the new Data Catalog',
+      sourcePage: 'catalog-list'
     }),
     thanksShowing && h(Modal, {
       onDismiss: () => setThanksShowing(false),
